refactor: change structure binding to values_view
