+++ src/graphics/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(${PROJECT_NAME}
         PRIVATE
+        vulkan/MeshNode.cpp
         vulkan/RenderableGLTF.cpp
         vulkan/vk_descriptors.cpp
         vulkan/vk_engine.cpp
